updated the backend

+++ Product_Store/backend/controllers/product.controller.js
@@ -33,12 +33,16 @@ export const deleteProduct = async (req, res) => {
     const {id} = req.params;
     // console.log("id: ", id);
 
+    if(!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: "Product not found." });
+    }
+
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "Product deleted successfully." });
     } catch (error) {
         console.log("Error in deleting product: ", error.message);
-        res.status(404).json({ success: false, message: "Product not found." });
+        res.status(500).json({ success: false, message: "Server error." });
     }
 }
